requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Dance[google]
python-dotenv